Debugged Training
Debugged training to the point where some of it should work more of the time now. The input layer is now trained (I didn't do that before by accident I swear). The other thing to note is that the XORText class outputs helpful dialog to help understand what the network is doing, so if you try to see what went wrong, I'd appreciate it.