Made Debugger and Helped Wes
I made a debugger much like the one found for Oponn. Unlike it, however, it has some added functionality. I also helped Wes in the sense that I made some of the variables we needed for the neural network and the heuristic available to him. I have not found a way for cards to be found out by any methods at all, but I am well on my way to do this.