Genome v5.5
Created a new tester class that should have some built-in debugging
things and allow you to specify what is to be tested. Added some more
functionality to the genome helper class. Fixed some things with the
epcoh class but still have to fix the ID issue. Fixed the multiple
genomes not firing by creating deep copies of link lists and neuron
lists. Created a print statement in the genome too.