Genome v6.1
Lots of changes to this one. I realized a few things. First, the
modified neural net actually doesn't talk to coved's. I forgot to remake
the network after modifying it. Luckily this was a relativly easy fix
after realizing what was wrong. I started work on the crossover function
and finished most of it. I will start testing the crossover tomorrow. I
also started work on the fitness function and believe I have a good way
to select genome based on their fitness. I hope by the end of the night
to have this completed by the end of the night.