Genome v6.2
So the add link function is what is causing errors and there is a bug to
be fixed inside the species class but other than that things work well
and I am able to produce multiple generations. things start to slow down
after 50 generations and I think it's because the innovation database so
I will look into that.  I can't begin training until Coved fixes the
link weights issue but once that is fixed training the genome to XOR
should be possible.