Genome v7.0
Probably one of the biggest visible sign of progress with this update. I
did make a mistake in the fitness function and the network does not
train to the XOR function but a different linearly seperable problem but
it runs through generations and modifies things and can "learn". I also
made improvement to the innovation database which fixed the slowness
issue with growing generations as I suspected. The improvement speeds
things up 10,000%. Yes 10,000% or 100x faster!! Anyway next up will be
fixing the addition of links, and editing the addition of neurons. After
that I will work to deactivate links and neurons.