Genome v9.0
This is the last genome before the final version. It kinda works. The
network outputs the same value regardless of the inputs which is really
confusing me and something I will need to look further into. After doing
this I am even more convinced RISK could never work. It took an hour to
get through 100 generations and it is fairly small. Speciation worked
though and it did eventually converge to the best answer it possibly
could given it ignored its inputs but that is not what we want it to do.